+++ modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-navigation-bars/src/main/java/com/liferay/fragment/collection/contributor/navigation/bars/internal/NavigationBarsFragmentCollectionContributor.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.fragment.collection.contributor.basic.component;
+package com.liferay.fragment.collection.contributor.navigation.bars.internal;
 
 import com.liferay.fragment.contributor.BaseFragmentCollectionContributor;
 import com.liferay.fragment.contributor.FragmentCollectionContributor;
+import com.liferay.fragment.contributor.PortletAliasRegistration;
 
 import javax.servlet.ServletContext;
 
@@ -23,18 +24,18 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Jürgen Kappler
+ * @author Eudaldo Alonso
  */
 @Component(
-	property = "fragment.collection.key=BASIC_COMPONENT",
+	property = "fragment.collection.key=NAVIGATION_BARS",
 	service = FragmentCollectionContributor.class
 )
-public class BasicComponentFragmentCollectionContributor
+public class NavigationBarsFragmentCollectionContributor
 	extends BaseFragmentCollectionContributor {
 
 	@Override
 	public String getFragmentCollectionKey() {
-		return "BASIC_COMPONENT";
+		return "NAVIGATION_BARS";
 	}
 
 	@Override
@@ -43,7 +44,12 @@ public class BasicComponentFragmentCollectionContributor
 	}
 
 	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.fragment.collection.contributor.basic.component)"
+		target = "(com.liferay.fragment.entry.processor.portlet.alias=nav)"
+	)
+	private PortletAliasRegistration _portletAliasRegistration;
+
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.fragment.collection.contributor.navigation.bars)"
 	)
 	private ServletContext _servletContext;
